Debugging GL calls.
These are generated by the gen_precalls.sh script.
Any OpenGL call prefixed with the letter 'd' get
a wrapper generated that calls a precall and postcall callback.
Basically a function gets executed before and after the GL call.
This is mostly useful for debugging purposes. The wrapper also prints
the location where the GL function has been called including the source
file and the line number. All the parameters that have been passed
into the function are printed for debugging purposes as well as the
return values of the functions that do return values.

It was primarily meant to be used to visualise the executing order
and the parameters passed into the GL functions for debugging purposes.
I am sure however there will be many other uses where this will come in
handy. It also allows the functionality to be selectively enabled
only for some GL calls by simply adding or removing the letter 'd'
prefix to any OpenGL function. Techically this could be expanded to work
with any function but it is not supported at the moment.

The script (gen_precalls.sh) is rather simple but it serves its purpose.