Prevent multiple inits.
Since Game::instance() can be called before the init sequence has
finished. It is possible to initalise multiple times if the has
intialised flag has been set to true just yet. That's why setting
the has initilised flag to true has to be done first. This also
allows the init functions to request the singleton instance now where
as before it would result in a multiple init bug.